requirements.txt
fastapi
uvicorn
pydantic
loguru
redis
python-jose
configparser
pytest
httpx
boto3
respx